requirements.txt
pillow
pyautogui
requests
discord-webhook
numpy
pytesseract
easyocr
ahk
mouse
keyboard
pynput
